libmbin 1.78 + AnimPose Classes
Started with entity parsing. For now one attachment is supported, we fetch it, then go through a component processing procedure. For now AnimPose and AnimData components are supported. AnimPose data has been exposed to the UI, I just need to update the pose on demand when the data changes. TODO: Do the same thing with the animations.