Fixed XAML issues with LibMBIN
I ended up compiling my own version of the library. Probably there was a conflict with the target architecture of the released lib. Also fixed any reference issues with all the other projects of the solution (Mainly QuickFont). Currently working on separating the jointDictionary from a scene, and moving it to locator objects.

The goal is to simply keep a local copy of the entire joint structure during scene import, which should be duplicated and stored in any locator object that has an animation component. This means that each locator will have its own version of the skeleton with a different pose and shit.

TODO: I don't trust the cloning functions at all. I simplified them a bit, but I need to work on them again in order to make sure that everything gets uniquely copied when generating a proc gen model.