Added UBO Support and organized the basic shaders
The whole pursuit was about implementing a better light system. It soon ended up with organizing all shader uniforms into structs, using UBOs for common perframe and permesh parameters. Because of the GLSL mess with explicit uniform locations in uniforms within a struct, I ended up loading all the active uniforms and keep a dictionary per shader with all the uniform locations.

I have to say that the rendering code is tremendously cleaner (although capturing the correct format of the C# structs to match the std140 GLSL versions was a nightmare), but performance seems pretty much the same, not to say it got a bit worse.