LOD Filtering, Overhaul of the GLSL preprocessing step and more
Primitive LOD filtering is now a thing, adopting properties of the entity files
In order to properly render transparent objects with whatever technique (sorted or unsorted) two things had to be implemented: 1) Output of the calculated light color to the gbuffer, therefore a new attachment was added. 2) Transparent objects need to be rendered in a second forward step and not using the deferred shader. This means that the same code has to be repeated twice, which will make it unmaintainable pretty much.

For this reason, I started an overhaul of the GLSL shader code preprocessing step, which keeps shader objects as separate entities, with separate file watchers for the included files, private compilation commands. This  will allow for separate shader versioning and directive definition, which helps tremendously on compiling separate branches of the same code. I am in the process of modifying shader source files to make this work back to normal. Then I'll test this with the deferred and forward lighting passes to see how this works.