Fixed (as much as I could for now)  transparent object rendering
Brought back to life back to front rendering of transparent objects
Removed all the BULLSHIT flags that i had as triggers for the transparent objects. Transparents are the ones that have TRANSPARENCY or CUTOUT (still not sure about that) and transparent scalar.
Cleaned up the shaders a bit


Also fixed the AABB calculation of mesh objects (for now the input value from the mbin attributes are transferred to the actual object)