AABB recalculation + checkpoint
Implemented a proper (expensive but correct) way of calculating the distance from a point to  an AABB. Sorting results look decent for now.

Encountered some problems regarding instanced data, spend the entire day trying to remember what the fuck was wrong. I finally remembered... skin matrices are shared amongst all instance of a mesh. This has to be split. I'm about to add a TBO per mesh, fit its size to the corresponding instance number of the mesh and use that in order to upload all instance skin data, without having problems with high instance counts and limited UBO sizes. On the other hand the corresponding data from mesh UBOs can be removed so this is also an extra gain.