A LOT of stuff once again
Disabled animation loading before the animation is used, in order to save time during scene loading. Actual animation data are loaded only when the animation is enabled for playback
Multiple shaders have been added to add support for bloom effect, gaussian blur and additive blending.
Added a shader compilation flag which for toggling saving log during shader compilation (seemed quite expensive) TODO: Expose that as a parameter in the settings file
Removed occlusion status per instance. Instances are repopulated on every frame depending on the visibility of the model. Therefore occlusion flag tests are also excluded from the shaders.
Small optimizations on camera frustum culling tests.
Small optimizations on the MathUtils class, for quicker access of Matrix4 elements
libPSARC added as a reference to the project. This is used in order to populate a list of all the game files through the .pak containers, and also to enable file loading directly from the paks. The actual list of all available scene files is a WIP to be exposed to the user under the open menu, but the file loaders are successfully fetching files from the paks. To save indexing time, a filemanifest file is saved when the app first starts, to keep the full mapping of files to their corresponding containers.
Positions are no longer included in the Gbuffer. actual pixel positions are calculated using the depth buffer and this also massively improved rendering because of much lighter bandwidth required.
Material shaders have been overhauled to work with material flag definitions. During material loading depending on the material flags the appropriate shader is compiled and stored to the app's memory for usage during rendering.
Slight improvements in the UI (added toggles for FXAA and BLOOM)
Dependency version updates
Added a very preliminary version of PBO buffers per texture for kinda asynchronous texture upload. Looks like its working good for now but I think there is still more into it.