Ray Casting attempt and more
After a shit ton of experiments with ray casting using the projection matrices, it just didn't work. I have no idea why that shit happens. But honestly I have absolutely no clue. Reverted to an alternative ray casting method even lighter that simply moves the camera front vector according to the mouse movement.

The problem is that even this is a valid method for casting rays from the camera position and maybe do particle shit and what not, its absolutely BS, for object selection, it just doesn't have the necessary precision. I need to revert back to the picking shader. The UI should be rendered to a different fbo. The scene has to be rendered to maybe 1/2 or even 1/4 of the main resolution using object IDs encoded as colors.  This should also happen to a different fbo as well. Picking involves reading from both fbos, first for checking the selected object, then for checking if the gizmo was selected. This is a WIP